chore(repo): refactor parser and cli-forge into multiple files
