+++ packages/parser/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/parser';
-export * from './lib/utils';
+export * from './lib/utils/case-transformations';
 export * from './lib/helpers';
+export * from './lib/option-types';
