+++ packages/parser/src/lib/option-types/boolean.ts
@@ -1,0 +1,5 @@
+import { CommonOptionConfig } from './common';
+
+export type BooleanOptionConfig<TCoerce = boolean, TChoices = TCoerce[]> = {
+  type: 'boolean';
+} & CommonOptionConfig<boolean, TCoerce, TChoices>;
