+++ packages/parser/src/lib/option-types/common.ts
@@ -1,0 +1,81 @@
+export type CommonOptionConfig<T, TCoerce = T, TChoices = T[]> = {
+  /**
+   * If set to true, the option will be treated as a positional argument.
+   */
+  positional?: boolean;
+
+  /**
+   * Provide an array of aliases for the option.
+   */
+  alias?: string[];
+
+  /**
+   * Provide an array of choices for the option. Values not in the array will throw an error.
+   */
+  choices?: TChoices | (() => TChoices);
+
+  /**
+   * Provide a default value for the option.
+   */
+  default?: T;
+
+  /**
+   * Provide a description for the option.
+   */
+  description?: string;
+
+  /**
+   * Provide a function to coerce the value of the option.
+   * @param value Value of the option
+   * @returns Coerced value of the option
+   */
+  coerce?: (value: T) => TCoerce;
+
+  /**
+   * Provide a function to validate the value of the option.
+   * @param value Coerced value of the option
+   * @returns If the value is valid, return true. If the value is invalid, return false or a string with an error message.
+   */
+  validate?: (value: TCoerce) => boolean | string;
+
+  /**
+   * If true, the option is required.
+   */
+  required?: boolean;
+
+  /**
+   * If set, the option will be populated from the environment variable `${env}_${optionName}`.
+   * If set to true, the environment variable will be `${optionName}`.
+   * If explicitly set to false, environment variable population will be disabled for this option.
+   */
+  env?:
+    | string
+    | boolean
+    | {
+        /**
+         * What key should the value be read from in process.env
+         */
+        key: string;
+
+        /**
+         * If set to false, ignore prefix provided by .env() call.
+         */
+        prefix?: boolean;
+      };
+
+  /**
+   * If set, the option will be marked as deprecated, with the provided message. This will not effect runtime behavior,
+   * but will be displayed in help output and generated docs.
+   */
+  deprecated?: string;
+
+  /**
+   * If true, the option will not be displayed in help output or generated docs.
+   */
+  hidden?: boolean;
+
+  /**
+   * Can be set to group options in help output and generated docs.
+   */
+  group?: string;
+};
