+++ packages/parser/src/lib/option-types/number.ts
@@ -1,0 +1,5 @@
+import { CommonOptionConfig } from './common';
+
+export type NumberOptionConfig<TCoerce = number, TChoices = TCoerce[]> = {
+  type: 'number';
+} & CommonOptionConfig<number, TCoerce, TChoices>;
