+++ packages/parser/src/lib/option-types/string.ts
@@ -1,0 +1,5 @@
+import { CommonOptionConfig } from './common';
+
+export type StringOptionConfig<TCoerce = string, TChoices = TCoerce[]> = {
+  type: 'string';
+} & CommonOptionConfig<string, TCoerce, TChoices>;
