+++ packages/parser/src/lib/parser.ts
@@ -1,163 +1,12 @@
+import { CommonOptionConfig } from './option-types/common';
 import { hideBin } from './helpers';
-import { fromDashedToCamelCase, getEnvKey } from './utils';
-
-type Flatten<T> = T extends Array<infer U> ? U : T;
-
-export type CommonOptionConfig<T, TCoerce = T, TChoices = T[]> = {
-  /**
-   * If set to true, the option will be treated as a positional argument.
-   */
-  positional?: boolean;
-
-  /**
-   * Provide an array of aliases for the option.
-   */
-  alias?: string[];
-
-  /**
-   * Provide an array of choices for the option. Values not in the array will throw an error.
-   */
-  choices?: TChoices | (() => TChoices);
-
-  /**
-   * Provide a default value for the option.
-   */
-  default?: T;
-
-  /**
-   * Provide a description for the option.
-   */
-  description?: string;
-
-  /**
-   * Provide a function to coerce the value of the option.
-   * @param value Value of the option
-   * @returns Coerced value of the option
-   */
-  coerce?: (value: T) => TCoerce;
-
-  /**
-   * Provide a function to validate the value of the option.
-   * @param value Coerced value of the option
-   * @returns If the value is valid, return true. If the value is invalid, return false or a string with an error message.
-   */
-  validate?: (value: TCoerce) => boolean | string;
-
-  /**
-   * If true, the option is required.
-   */
-  required?: boolean;
-
-  /**
-   * If set, the option will be populated from the environment variable `${env}_${optionName}`.
-   * If set to true, the environment variable will be `${optionName}`.
-   * If explicitly set to false, environment variable population will be disabled for this option.
-   */
-  env?:
-    | string
-    | boolean
-    | {
-        /**
-         * What key should the value be read from in process.env
-         */
-        key: string;
-
-        /**
-         * If set to false, ignore prefix provided by .env() call.
-         */
-        prefix?: boolean;
-      };
-
-  /**
-   * If set, the option will be marked as deprecated, with the provided message. This will not effect runtime behavior,
-   * but will be displayed in help output and generated docs.
-   */
-  deprecated?: string;
-
-  /**
-   * If true, the option will not be displayed in help output or generated docs.
-   */
-  hidden?: boolean;
-
-  /**
-   * Can be set to group options in help output and generated docs.
-   */
-  group?: string;
-};
-
-export type StringOptionConfig<TCoerce = string, TChoices = TCoerce[]> = {
-  type: 'string';
-} & CommonOptionConfig<string, TCoerce, TChoices>;
-
-export type NumberOptionConfig<TCoerce = number, TChoices = TCoerce[]> = {
-  type: 'number';
-} & CommonOptionConfig<number, TCoerce, TChoices>;
-
-export type BooleanOptionConfig<TCoerce = boolean, TChoices = TCoerce[]> = {
-  type: 'boolean';
-} & CommonOptionConfig<boolean, TCoerce, TChoices>;
-
-export type StringArrayOptionConfig<TCoerce = string, TChoices = TCoerce[]> = {
-  type: 'array';
-  items: 'string';
-} & CommonOptionConfig<string[], TCoerce, TChoices>;
-
-export type NumberArrayOptionConfig<TCoerce = number, TChoices = number[]> = {
-  type: 'array';
-  items: 'number';
-} & CommonOptionConfig<number[], TCoerce, TChoices>;
-
-export interface ObjectOptionConfig<
-  TCoerce = Record<string, string>,
-  TProperties extends {
-    [key: string]: Readonly<OptionConfig>;
-  } = Record<string, any>
-> extends Omit<CommonOptionConfig<Record<string, string>, TCoerce>, 'choices'> {
-  type: 'object';
-  /** */
-  properties: TProperties;
-  additionalProperties?: false | 'string' | 'number' | 'boolean';
-}
-
-/**
- * Configuration for array options. Arrays are parsed from
- * comma separated, space separated, or multiple values.
- *
- * e.g. `--foo a b c`, `--foo a,b,c`, or `--foo a --foo b --foo c`
- */
-export type ArrayOptionConfig<
-  TCoerce = string | number,
-  TChoices = TCoerce[]
-> =
-  | StringArrayOptionConfig<TCoerce, TChoices>
-  | NumberArrayOptionConfig<TCoerce, TChoices>;
-
-/**
- * Configures an option for the parser. See subtypes for more information.
- * - {@link StringOptionConfig}
- * - {@link NumberOptionConfig}
- * - {@link ArrayOptionConfig}
- * - {@link BooleanOptionConfig}
- *
- * @typeParam TCoerce The return type of the `coerce` function if provided.
- */
-export type OptionConfig<
-  TCoerce = any,
-  TChoices = any[],
-  TObjectProps extends Record<string, OptionConfig> = Record<string, any>
-> =
-  | StringOptionConfig<TCoerce, TChoices>
-  | NumberOptionConfig<TCoerce, TChoices>
-  | ArrayOptionConfig<TCoerce, TChoices>
-  | BooleanOptionConfig<TCoerce, TChoices>
-  | ObjectOptionConfig<TCoerce, TObjectProps>;
-
-export type InternalOptionConfig = OptionConfig & {
-  key: string;
-  position?: number;
-};
-
-export type Internal<T extends OptionConfig> = T & InternalOptionConfig;
+import { OptionConfigToType } from './option-types/option-config-to-type';
+import { fromDashedToCamelCase, getEnvKey } from './utils/case-transformations';
+import { InternalOptionConfig, OptionConfig } from './option-types';
+import { parserMap } from './parsers/parser-map';
+import { NoValueError, Parser, ParserContext } from './parsers/typings';
+import { getConfiguredOptionKey } from './utils/get-configured-key';
+import { isFlag, isNextFlag, readArgKeys } from './utils/flags';
 
 /**
  * Base type for parsed arguments.
@@ -680,319 +529,7 @@ function validateOption<T>(optionConfig: InternalOptionConfig, value: T) {
   }
 }
 
-function getConfiguredOptionKey<T extends ParsedArgs>(
-  key: string,
-  configuredOptions: Partial<Record<keyof T, OptionConfig>>
-): keyof T | undefined {
-  if (key in configuredOptions) {
-    return key as keyof T;
-  }
-
-  function normalizeNegatedBooleanKey(key: string) {
-    if (
-      key.startsWith('no') &&
-      key.length > 2 &&
-      key[2] === key[2].toUpperCase()
-    ) {
-      const stripped = key.slice(2);
-      return [stripped[0].toLowerCase(), stripped.slice(1)].join('');
-    }
-    return key;
-  }
-
-  function normalizeObjectOptionKey(key: string) {
-    if (key.includes('.')) {
-      return key.split('.')[0];
-    }
-    return key;
-  }
-
-  // Handles booleans passed as `--no-foo`
-  const normalizedKey = normalizeObjectOptionKey(
-    normalizeNegatedBooleanKey(key)
-  );
-
-  if (normalizedKey in configuredOptions) {
-    return normalizedKey as keyof T;
-  }
-
-  for (const configuredKey in configuredOptions) {
-    const config = configuredOptions[configuredKey];
-    if (config?.alias?.includes(key)) {
-      return configuredKey as keyof T;
-    }
-    // Handles negated booleans that are aliased
-    if (config?.type === 'boolean' && config.alias?.includes(normalizedKey)) {
-      return configuredKey as keyof T;
-    }
-  }
-  return undefined;
-}
-
-function isNextFlag(str: string) {
-  return str.startsWith('--') || str.startsWith('-');
-}
-
-const booleanParser: Parser<Internal<BooleanOptionConfig>> = ({
-  tokens,
-  providedFlag,
-}) => {
-  const negated = providedFlag?.startsWith('--no-');
-  const val = tokens.shift();
-  const parsed = (() => {
-    if (val === undefined) {
-      return true;
-    }
-    if (isNextFlag(val)) {
-      tokens.unshift(val);
-      return true;
-    }
-    if (val === 'false') {
-      return false;
-    }
-    return true;
-  })();
-  return negated ? !parsed : parsed;
-};
-
-export class NoValueError extends Error {
-  constructor() {
-    super('Expected a value');
-  }
-}
-
-const stringParser: Parser<Internal<StringOptionConfig>> = ({ tokens }) => {
-  const val = tokens.shift();
-  if (val === undefined) {
-    throw new NoValueError();
-  }
-  if (isNextFlag(val)) {
-    tokens.unshift(val);
-    throw new NoValueError();
-  }
-  return val;
-};
-
-const numberParser: Parser<Internal<NumberOptionConfig>> = ({ tokens }) => {
-  const val = tokens.shift();
-  if (val === undefined) {
-    throw new NoValueError();
-  }
-  if (isNextFlag(val)) {
-    tokens.unshift(val);
-    throw new NoValueError();
-  }
-  return Number(val);
-};
-
-const quotePairs = {
-  '"': '"',
-  "'": "'",
-  '`': '`',
-} as const;
-
-const csvParser = (str: string) => {
-  const collected = [];
-  let val = '';
-  let inQuote: keyof typeof quotePairs | false = false;
-  for (const char of str) {
-    if (inQuote) {
-      if (char === quotePairs[inQuote]) {
-        inQuote = false;
-      } else {
-        val += char;
-      }
-    } else {
-      if (char in quotePairs) {
-        inQuote = char as keyof typeof quotePairs;
-      } else if (char === ',') {
-        collected.push(val);
-        val = '';
-      } else {
-        val += char;
-      }
-    }
-  }
-  collected.push(val);
-  return collected;
-};
-
-const arrayParser: Parser<Internal<ArrayOptionConfig>> = <
-  T extends string | number
->({
-  config,
-  tokens,
-  current,
-}: ParserContext<Internal<ArrayOptionConfig>>) => {
-  const coerce =
-    config.items === 'string'
-      ? (s: string) => s as T
-      : (s: string) => Number(s) as T;
-  let val = tokens.shift();
-  if (val && val.includes(',')) {
-    return csvParser(val).map(coerce);
-  }
-  const collected: string[] = [];
-  while (val) {
-    if (isNextFlag(val)) {
-      tokens.unshift(val);
-      break;
-    }
-    collected.push(val);
-    val = tokens.shift();
-  }
-  const coerced = collected.map(coerce);
-  return current ? current.concat(coerced) : coerced;
-};
-
-const objectParser: Parser<Internal<ObjectOptionConfig>> = ({
-  tokens,
-  config,
-  providedFlag,
-  current,
-}) => {
-  current ??= {};
-  // `providedFlag` is the flag that was used on the cli.
-  // it will look like `env.foo.bar`, or `env.foo`. for dot notation
-  //
-  // We don't care about the first part, as the base parser has already matched the flag.
-  const parts = providedFlag?.split('.').slice(1);
-  if (!parts?.length) {
-    throw new Error(
-      `${config.key} is configured as an object, but no properties were provided. Pass properties like so: --${config.key}.foo bar --${config.key}.baz qux`
-    );
-  }
-  const { config: propConfig, readValue, setValue } = parsePath(parts);
-  const currentValue = readValue();
-  const parsedValue = tryParseValue(parserMap[propConfig.type], {
-    config: {
-      ...propConfig,
-      key: `${config.key}.${parts.join('.')}`,
-    },
-    tokens,
-    current: currentValue,
-    providedFlag: providedFlag,
-  });
-  setValue(parsedValue);
-
-  return current;
-
-  function parsePath(parts: string[]): {
-    readValue(): any;
-    setValue(v: any): void;
-    config: OptionConfig;
-  } {
-    const propParts = [...parts];
-    let currentObject = current;
-    let currentKey = propParts.shift();
-    let currentValue: any;
-    let currentConfig: OptionConfig = config;
-    let last: string;
-    // eslint-disable-next-line no-constant-condition
-    while (true) {
-      if (!currentKey) {
-        return {
-          readValue() {
-            return currentValue;
-          },
-          setValue(v) {
-            currentObject[last] = v;
-          },
-          config: currentConfig,
-        };
-      }
-      const nextKey = propParts.shift();
-      currentValue = currentObject[currentKey];
-      if (nextKey) {
-        if (currentValue && typeof currentValue !== 'object') {
-          throw new Error(
-            `Expected ${currentKey} to be an object, but found ${typeof currentValue}`
-          );
-        } else {
-          currentObject[currentKey] ??= {};
-        }
-      }
-      last = currentKey;
-      currentKey = nextKey;
-      if (nextKey) {
-        currentObject = currentObject[last];
-      }
-      const c = (currentConfig as ObjectOptionConfig).properties[last];
-      if (!c) {
-        if (
-          'additionalProperties' in currentConfig &&
-          currentConfig.additionalProperties
-        ) {
-          currentConfig = {
-            type: currentConfig.additionalProperties,
-          };
-        } else {
-          throw new Error(
-            `No configuration found for ${last} in ${config.key}`
-          );
-        }
-      } else {
-        currentConfig = c;
-      }
-    }
-  }
-
-  // function readValue(propParts: string[]) {
-  //   const pathParts = [...propParts];
-  //   let currentValue = current;
-  //   let currentKey = pathParts.shift();
-  //   while (currentKey) {
-  //     if (typeof currentValue !== 'object' && currentValue !== undefined) {
-  //       throw new Error(
-  //         `Expected ${currentKey} to be an object, but found ${typeof currentValue}`
-  //       );
-  //     } else if (!(currentKey in currentValue)) {
-  //       return undefined;
-  //     } else {
-  //       currentValue = currentValue[currentKey];
-  //       currentKey = pathParts.shift();
-  //     }
-  //   }
-  // }
-
-  // function readOptionConfig(propParts: string[]): InternalOptionConfig {
-  //   const pathParts = [...propParts];
-  //   let currentConfig = config;
-  //   let currentKey = pathParts.shift();
-  //   while (currentKey) {
-  //     if (!(currentKey in currentConfig.properties)) {
-  //       if (!currentConfig.additionalProperties) {
-  //         throw new Error(
-  //           `No configuration found for ${currentKey} in ${config.key}`
-  //         );
-  //       } else {
-  //         return {
-  //           type: currentConfig.additionalProperties,
-  //           key: 'additionalProperties',
-  //         };
-  //       }
-  //     }
-  //     currentConfig = currentConfig.properties[currentKey];
-  //     currentKey = pathParts.shift();
-  //   }
-  //   return currentConfig;
-  // }
-
-  // function setValue(propParts: string[]) {}
-};
-
-type ParserContext<TConfig extends OptionConfig> = {
-  config: TConfig;
-  tokens: string[];
-  current?: any;
-  providedFlag?: string;
-};
-
-type Parser<TConfig extends OptionConfig, T = any> = (
-  input: ParserContext<TConfig>
-) => T;
-
-function tryParseValue(
+export function tryParseValue(
   parser: Parser<InternalOptionConfig>,
   input: ParserContext<InternalOptionConfig>
 ) {
@@ -1015,33 +552,6 @@ function tryParseValue(
   }
 }
 
-const parserMap: Record<string, Parser<any>> = {
-  string: stringParser,
-  number: numberParser,
-  boolean: booleanParser,
-  array: arrayParser,
-  object: objectParser,
-};
-
-function isFlag(str: string): str is `-${string}` {
-  return str.startsWith('-');
-}
-
-function readArgKeys(str: `-${string}`): string[] {
-  // Long flags (e.g. --foo)
-  if (str.startsWith('--')) {
-    const key = str.slice(2);
-    if (key.includes('-')) {
-      return [fromDashedToCamelCase(key)];
-    }
-    return [str.slice(2)];
-    // Short flag combinations (e.g. -xvf)
-  } else if (str.startsWith('-')) {
-    return str.slice(1).split('');
-  }
-  throw new Error(`Invalid flag ${str}`);
-}
-
 export class ValidationFailedError<T> extends AggregateError {
   constructor(
     errors: Error[],
@@ -1051,45 +561,3 @@ export class ValidationFailedError<T> extends AggregateError {
     super(errors, message);
   }
 }
-
-export type OptionConfigToType<TOptionConfig extends OptionConfig> =
-  InferTChoice<TOptionConfig> extends [never]
-    ? TOptionConfig['coerce'] extends (s: any) => any
-      ? ReturnType<TOptionConfig['coerce']>
-      : {
-          string: string;
-          number: number;
-          boolean: boolean;
-          array: ArrayItems<TOptionConfig>[];
-          object: TOptionConfig extends ObjectOptionConfig
-            ? ResolveTProperties<TOptionConfig['properties']> &
-                AdditionalProperties<TOptionConfig>
-            : never;
-        }[TOptionConfig['type']]
-    : InferTChoice<TOptionConfig>;
-
-type ArrayItems<TOptionConfig extends OptionConfig> =
-  TOptionConfig extends ArrayOptionConfig<string | number>
-    ? TOptionConfig['items'] extends 'string'
-      ? string
-      : number
-    : never;
-
-type AdditionalProperties<TOptionConfig extends ObjectOptionConfig> =
-  TOptionConfig['additionalProperties'] extends 'string'
-    ? Record<string, string>
-    : TOptionConfig['additionalProperties'] extends 'number'
-    ? Record<string, number>
-    : TOptionConfig['additionalProperties'] extends 'boolean'
-    ? Record<string, boolean>
-    : Record<string, never>;
-
-type ResolveTProperties<TProperties extends Record<string, OptionConfig>> = {
-  [key in keyof TProperties]: OptionConfigToType<TProperties[key]>;
-};
-
-type InferTChoice<TOptionConfig> = 'choices' extends keyof TOptionConfig
-  ? TOptionConfig['choices'] extends Array<infer TChoice>
-    ? TChoice
-    : never
-  : never;
